[ucsim] Apply [patches:#421]
git-svn-id: http://svn.code.sourceforge.net/p/sdcc/code/trunk/sdcc@12930 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ sim/ucsim/sim.src/uccl.h
@@ -68,7 +68,7 @@ public:
   virtual ~cl_ticker(void);
 
   enum ticker_type get_type(void) { return type; }
-  double get_ticks(void) { return ticks; }
+  unsigned long get_ticks(void) { return ticks; }
   double get_rtime(void) { return rtime; }
 
   void set(int nr, double time) { ticks = nr; rtime = time; }
